poprawki

+++ src/main/java/psk/projects/dating_portal/auth/UserRepository.java
@@ -2,7 +2,7 @@ package psk.projects.dating_portal.auth;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<AppUser, Integer> {
+public interface UserRepository extends CrudRepository<AppUser, Long> {
     long countByLogin(String login);
     boolean existsByLogin(String login);
     AppUser findByLogin(String login);
